requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.10
openai==1.6.1
chromadb==0.4.22
tiktoken==0.5.2
requests==2.31.0
gunicorn==21.2.0